+++ Makefile
@@ -485,16 +485,8 @@ main-build-dockerized: build-volumes
 
 .PHONY: main-build-nodeps
 main-build-nodeps:
-	$(GOBUILD) \
-		central \
-		compliance/cmd/compliance \
-		config-controller \
-		migrator \
-		operator/cmd \
-		sensor/admission-control \
-		sensor/kubernetes \
-		sensor/upgrader \
-		compliance/virtualmachines/roxagent
+	$(GOBUILD) central
+	$(GOBUILD) operator/cmd
 	mv bin/linux_$(GOARCH)/cmd bin/linux_$(GOARCH)/stackrox-operator
 ifndef CI
 	CGO_ENABLED=0 $(GOBUILD) roxctl
