+++ scripts/check-image-version.sh
@@ -3,7 +3,7 @@
 tmpfile="$(mktemp)"
 trap 'rm -f "${tmpfile}"' EXIT
 
-git grep -E -o -h '(stackrox|scanner)-(build|test)-[0-9]+\.[0-9]+\.[0-9]' | grep -E -o '[0-9]+\.[0-9]+\.[0-9]' | sort -u >"$tmpfile"
+git grep -E -o -h '(stackrox|scanner)-(build|test)-[0-9]+\.[0-9]+\.[0-9].*' | grep -E -o '[0-9]+\.[0-9]+\.[0-9]' | sort -u >"$tmpfile"
 
 if [[ "$( wc -w < "$tmpfile" )" -eq 1 ]]
  then
