+++ sensor/admission-control/app/certs.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"os"
